pyproject.toml
[project]
name = "balanceflo-assignment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ipykernel>=6.30.1",
    "matplotlib>=3.10.6",
    "numpy>=2.3.2",
    "opencv-python>=4.11.0.86",
    "opencv-python-headless>=4.11.0.86",
    "pandas>=2.3.2",
    "pillow>=11.3.0",
    "plotly>=6.3.0",
    "scikit-image>=0.25.2",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "streamlit>=1.49.1",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
]
